LayoutTests/accessibility/mac/resolved-text-editing.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>

<div contenteditable id=div></div>

<script>
let output = "This tests that we can retrieve resolved editing styles from the editable text area.\n\n";

if (window.accessibilityController) {
    window.jsTestIsAsync = true;

    var div = document.getElementById("div");
    div.focus();

    textInputController.insertText("a");
    document.execCommand("italic", false)
    document.execCommand("FontSize", false, 6)

    var axDiv = accessibilityController.accessibleElementById("div");
    axDiv.attributeValueAsync("AXResolvedEditingStyles", function(value) {
        output += value;

        debug(output);
        finishJSTest();
    });
}
</script>
</body>
</html>
